1. added functionality to delete the corresponding blocks to an unauthorized user, as:  - added question stimulus controller  - refactoring the corresponding views
