+++ app/javascript/controllers/index.js
@@ -6,5 +6,8 @@ eagerLoadControllersFrom("controllers", application)
 import AuthorizeController from "./authorize_controller"
 application.register("authorize", AuthorizeController)
 
-import AnswersController from "./answer_controllers/answers_controller"
+import AnswersController from "./answers_controllers/answers_controller"
 application.register("answers", AnswersController)
+
+import QuestionsController from "./questions_controllers/questions_controller"
+application.register("questions", QuestionsController)
