+++ app/javascript/controllers/questions_controllers/questions_controller.js
@@ -1,0 +1,9 @@
+import AuthorizeController from "../authorize_controller"
+
+export default class extends AuthorizeController {
+    static targets = ["question"]
+
+    questionTargetConnected(element) {
+        this.processElement(element);
+    }
+}
